[fix] code review comment things fix
