+++ src/EnrollRestaurant/EnrollRestaurant.tsx
@@ -2,7 +2,6 @@ import { ThemeProvider } from "@emotion/react";
 import {
   Button,
   Container,
-  createTheme,
   Paper,
   Step,
   StepLabel,
@@ -18,17 +17,6 @@ import store from "./store/store";
 import DetailInfo from "./DetailInfo";
 import SelectShare from "./SelectShare";
 import Confirm from "./Confirm";
-
-const theme = createTheme({
-  palette: {
-    secondary: {
-      main: "#F8E9A1",
-    },
-    primary: {
-      main: "#24305E",
-    },
-  },
-});
 
 const steps = ["기본정보", "종류", "상세정보", "공유목록 선택", "확인"];
 
@@ -58,53 +46,46 @@ const EnrollRestaurant: FC = observer(() => {
   };
 
   return (
-    <ThemeProvider theme={theme}>
-      <Container component="main" maxWidth="sm" sx={{ mb: 4 }}>
-        <Paper variant="outlined" sx={{ p: { xs: 2, md: 3 } }}>
-          <Typography component="h1" variant="h4" align="center">
-            맛집 등록
-          </Typography>
-          <Stepper
-            activeStep={activeStep}
-            alternativeLabel
-            sx={{ pb: 5, pt: 3 }}
-          >
-            {steps.map((label) => (
-              <Step key={label}>
-                <StepLabel>{label}</StepLabel>
-              </Step>
-            ))}
-          </Stepper>
+    <Container component="main" maxWidth="sm" sx={{ mb: 4 }}>
+      <Paper variant="outlined" sx={{ p: { xs: 2, md: 3 } }}>
+        <Typography component="h1" variant="h4" align="center">
+          맛집 등록
+        </Typography>
+        <Stepper activeStep={activeStep} alternativeLabel sx={{ pb: 5, pt: 3 }}>
+          {steps.map((label) => (
+            <Step key={label}>
+              <StepLabel>{label}</StepLabel>
+            </Step>
+          ))}
+        </Stepper>
 
-          {getStepContent(activeStep)}
+        {getStepContent(activeStep)}
 
-          
-            {activeStep < steps.length - 1 ? (
-                <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
-                    <Button
-                        variant="contained"
-                        color="primary"
-                        sx={{ mt: 3, ml: 1 }}
-                        onClick={handleNext}
-                    >
-                        Next
-                    </Button>
-                </Box>
-            ) : (
-                <Box sx={{ display: "flex", justifyContent: 'center'}}>
-                    <Button
-                        variant="contained"
-                        color="primary"
-                        sx={{ mt: 3, ml: 1 }}
-                        onClick={handleNext}
-                    >
-                        완료
-                    </Button>
-                </Box>
-            )}
-        </Paper>
-      </Container>
-    </ThemeProvider>
+        {activeStep < steps.length - 1 ? (
+          <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
+            <Button
+              variant="contained"
+              color="primary"
+              sx={{ mt: 3, ml: 1 }}
+              onClick={handleNext}
+            >
+              Next
+            </Button>
+          </Box>
+        ) : (
+          <Box sx={{ display: "flex", justifyContent: "center" }}>
+            <Button
+              variant="contained"
+              color="primary"
+              sx={{ mt: 3, ml: 1 }}
+              onClick={handleNext}
+            >
+              완료
+            </Button>
+          </Box>
+        )}
+      </Paper>
+    </Container>
   );
 });
